requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
yfinance>=0.2.36
fredapi>=0.5.1
tenacity>=8.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2
scipy>=1.12.0
pyarrow>=15.0.0
feedparser>=6.0.10
